test/aiTest.py
import openai
from Token import OPENAITOKEN

openai.api_key = OPENAITOKEN

response = openai.Completion.create(
    model="text-davinci-003",
    prompt="講個笑話來聽聽",
    max_tokens=128,
    temperature=0.5,
)

completed_text = response["choices"][0]["text"]
print(completed_text)