requirement.txt
discord.py==1.7.3
pytube==11.0.1
youtube-dl==2022.2.10
PyNaCl==1.4.0